fix fix Residual Filter containing wrong info
